--- src/App.tsx
@@ -1,12 +1,0 @@
-import { BrowserRouter } from 'react-router-dom';
-import Home from './main_component/home';
-
-function App() {
-  return (
-    <BrowserRouter>
-    <Home />
-    </BrowserRouter>
-  );
-}
-
-export default App;
