+++ src/app/App.tsx
@@ -1,0 +1,9 @@
+import Home from '../pages/home';
+
+function App() {
+  return (
+    <Home />
+  );
+}
+
+export default App;
