+++ src/css/Home/Accouncement/Inventory.module.css
