+++ src/features/main_component/Announcement/Announcement.tsx
