+++ src/features/main_component/Announcement/Inventory.tsx
