+++ src/features/main_component/Class/Card.tsx
