+++ src/features/main_component/Class/GoToClass.tsx
