+++ src/features/main_component/Class/MiniCard.tsx
