+++ src/features/main_component/Footer/TeamFooter.tsx
