+++ src/features/main_component/Other/OtherCard.tsx
